#1363 move XATransactionDataSourceWrapper => transaction.xa.manager.
